fix: 돈이 Int 타입의 한계를 넘어버리기에, Long 타입으로 변경

+++ src/main/kotlin/domain/money/Money.kt
@@ -1,6 +1,6 @@
 package domain.money
 
-data class Money(val value: Int) : Comparable<Money> {
+data class Money(val value: Long) : Comparable<Money> {
     operator fun plus(other: Money): Money {
         return Money(this.value + other.value)
     }
@@ -13,7 +13,7 @@ data class Money(val value: Int) : Comparable<Money> {
         return this.value.compareTo(other.value)
     }
 
-    fun dividedBy(other: Money): Int {
+    fun dividedBy(other: Money): Long {
         return this.value / other.value
     }
