+++ src/main/kotlin/view/BuyingInput.kt
@@ -2,8 +2,8 @@ package view
 
 import domain.money.Money
 
-data class BuyingInput(val amount: Int) {
-    constructor(amount: String) : this(amount.toInt())
+data class BuyingInput(val amount: Long) {
+    constructor(amount: String) : this(amount.toLong())
 
     fun toMoney(): Money = Money(amount)
 }
